moving functions to talklib package
we're moving the send_sms and send_call functions to our main Python package, talklib. this way we can reuse this functionality in other scripts.

these changes simplify this script, but the obvious drawback is the script now has a dependency (talklib). however, talklib should be installed on any/all TL PCs regardless.