attempt to kill process
ncat is still occasionally freezing on us. thus, when we try to send the next message, ncat is hung up, and it refuses to send. this is an attempt to force the shell to close once the attempt has been made.

TODO split this into functions and what not. haven't done that because of the -u flag for sending via UDP to BrightSign.