NO MORE NCAT!!!!!!!!!!!!!!!!
finally, at long last, after a month+ of trying to get this to work......it seems we have a working, pure python way of sending Now Playing metadata to our Telos unit (which forwards it to Icecast) AND the BrightSign unit.

it uses Python's socket library, part of the standard library. I had tried experimenting with this before but could not get it working. it turns out the problem had to do with the metadata filter we were using on the Telos unit. now we switched to this more explicit filter (XML Parser Sample), which gives easier formatting that won't confuse Python or the Telos unit.

separately, we're using UDP to send to BrightSign (only accepts UDP, weirdly), and that doesn't require filtering/parsing.

YAAAAAAAAYYYYY!!!!!! :)