add HealthHeart
The RSS feed for Health in a Heartbeat has become too unreliable to use. At the beginning of each month, the feed stops updating and often doesn't start again until midway through the month. We end up having to download the files manually so often that it makes more sense to just download everything at the beginning of the month and run this script.

cheers